[Project] Move ratelimit parsing stuff to a common library
